Align DEV_SEED_STUDIO demo user with /api/auth/dev and extract helper
